feat: Add -> operator which get property value from vertex
It is preferable to use `.` to get a property value from the given
vertex. But it is hard to use `.` because the scanner treats `.` as a
character, not an operator, and indirection rule uses `.` for the same
purpose. So, we add it as a vertex -> operator first.
The vertex -> operator actually calls the function for jsonb -> operator
with a Jsonb in the vertex and a property key as arguments.

(cherry picked from commit a307376b444de662ad8a2cd0a932063acf9f6d85)