feat(cypher): Implement pattern grammar
Now, the parser can parse Cypher's pattern. However, there are some
workarounds due to the behavior of the scanner. First, we cannot parse
a very simple relationship `--` because the scanner treats it as an
inline comment. So, current implementation only can support it through
`-[]-`. Also, right direction `->` and `|` which connects multiple
relationship type constraints are treated as an operator. So, we use
`Op` token in the rule instead, and if the parser detects a parsed
token other than `->` and `|`, then it reports a syntax error.