feat: Add pg_statsinfo-3.1.2 into contrib
We use pg_statsinfo for monitoring Agens Graph server.
It is included in contrib, so to build it use 'make world' and 'make install-world'.
In order to run pg_statsinfo's agens in server,
postgres.conf should be modified. Please refer to the documents of pg_statsinfo.
The default postgres.conf will be modified to run pg_statsinfo.
repoter module in pg_statsinfo is excluded because it is for web-UI.
Some symbols like INIT_CRC32 are modified to the current version of Agens Graph.