feat: Support `EXISTS()` and `SIZE()`
These expressions get a pattern (not path) as their argument and
return the existance/number of the pattern.

BREAKING CHANGE:

Since `SIZE()` is implemented as an expression, if a user add a new
function named "size" and call it, it will never be called.

(cherry picked from commit 742264186437f70929822acbb5573f241fe8821c)