feat: Make use of property index for property constraints
Given a property constraint like `{name: {first: 'Tom'}, age: 30}`,
it will be transformed to as below.
```
WHERE properties #>> '{name,first}'::text[] AND
      properties #>> '{age}'::text[]
```
And if one of the target label and its sub-labels has a GIN index on
the `properties` column, it will add the following qualification.
`properties @> jsonb_build_object(...)`