fix: Duplicate variable when transforming SET clause
After future vertices are resolved, a new RTE is added to the rtable
list. When transforming expressions which uses variables from previous
clauses, there might be multiple visible variables of the same name in
the RTE of the previous clause and the newly added RTE. So, hide the
variables from the previous clause on this situation.

Code and commit message cleanup by Junseok Yang.

Closes #180