feat: Implement map literal
If all property values in a map literal are constant, the map literal
will be pre-evaluated to make it a single constant.

Here is an example of map literal.

```
{
    i: 7,
    n: 7.0,
    s: '"Hello\nAgensGraph\"',
    t: true,
    f: false,
    '0': null,
    '\n': 'AS-IS',
    o: {
        v: 'bitnine',
        o: {
            v: 'bitnine'
        }
    }
}
```

And the following is the evaluated result. Properties that have null as
their value are omitted.

```
{
    "f": false,
    "i": 7,
    "n": 7.0,
    "o": {
        "o": {
            "v": "bitnine"
        },
        "v": "bitnine"
    },
    "s": "\"Hello\nAgensGraph\"",
    "t": true,
    "\\n": "AS-IS"
}
```

(cherry picked from commit f2e6bdf40a61a4339232b6a2b8953ccec7542440)