feat: Implement comparison and boolean operators
The implementation of comparison operators (=, <>, <, >, <=, >=) uses
existing `jsonb` comparison functions. This has pros and cons. It
allows optimizer to make use of indexes. And the implementation is
relatively easy. However, those functions return `bool` value. It means
if users want to use the value as a literal value, type conversion from
`boot` to `jsonb` is required. So, implicit type conversion between
`jsonb` and `bool` is introduced.

The implementation of boolean operators (OR, AND, NOT) also uses
existing boolean expression implementation. It has similar pros and
cons to that of comparison operators.

(cherry picked from commit 19706eac004699d00db4c47821d7d6fe6735b496)