feat: Implement property access
* Accessing non-exisiting properties returns null.
* `o.null`, `o.'null'`, `o[null]`, `o[NULL]`, and `o['null']` are equal.
* `o.true`, `o.'true'`, `o[true]`, `o[TRUE]`, and `o['true']` are equal.
* Property access on null is null. (e.g. `null.p`)
* Number, map, and list are invalid for the key value.

Let's assume that there is a map `m` with `{p1: 0, 'p\n2': {p3: 1}}`.

Both `m.p1` and `m['p' + 1]` will return `0`.

`m.'p\n2'` will return `{p3: 1}`. However, `m['p\n2']` will return null.
This is because `'p\n2'` in the definition of `m` is equal to that in
`m.'p\n2'` (a string which has 4 characters) but is different from that
in `m['p\n2']` (a string which has 2 characters and a newline character
in the middle). The `.` operator takes an identifier and use it AS-IS.
On the other hand, the `[]` operator takes an expression. This means
escaped characters in the string are processed before using it.

Users may use these two operators consecutively. For example,

`m.p1['p\\n2'].'p3'` will return `1`.

(cherry picked from commit 9f4196af6a715f17b1f0eda3ad8a2150b470ccf1)