fix: Handle parameters with `UNKNOWNOID` type
Current implementation of `transformParamRef()` in Cypher expression
calls `to_jsonb()` over parameters to convert those parameters to
`jsonb` value.

PostgreSQL cannot decide which function to use for the conversion when
the type of a parameter is `UNKNOWNOID` because there are three
`to_jsonb()` functions. (One for `jsonb` and others for `vertex` and
`edge`.)

Now, `coerce_type()` is called instead of `to_jsonb()` only if the type
of the parameter is `UNKNOWNOID`. This call will change the type to
`JSONBOID`.

When binding the actual parameter value to the parameter, `jsonb_in()`
or `jsonb_recv()` will be called with the value.