fix: Handle `JsonbValue.val.string.val` properly
It does not end with `'\0'` but `ExecEvalCypherTypeCast()` does not
care this properly. For example, `RETURN ('1' + '2')::text;` returns a
result like `"12~\x7F@"`. And we can save 1 byte if a C-string is
stored in it. This is applied to `jsonb_add()`.

(cherry picked from commit 3afa1cc4f69c19c246789bd90303606a49359b5e)