fix: Cypher `substring` function uses wrong start index (AG246)
This patch fixes the cypher version of `substring` so that the
starting index is 0. It does this by adding the operator
function `+` with the constant 1 to the starting index expression.
This is then passed to the SQL version of `substring` which starts
at 1. Cypher strings start at 0 while SQLs start at 1.

Additionally, it corrects a typo in the name of the function
`fucn_get_best_args` to `func_get_best_args`. It is a static
function, so all local occurances where changed (3).

The following functions were modified -

* `fucn_get_best_args` - the name.

* `func_get_best_candidate` added in code to remove the jsonb
  `substring` function from the raw candidates iff it is the
  `substring` function being processed.

* `proprocess_func_args` added in code to construct an `AEXPR_OP`
   to add 1 to the `substring` start parameter.

Also added the following files to regression tests -

* src/test/regress/sql/cypher_substring.sql (sql test script)

* src/test/regress/expected/cypher_substring.out (expected output)

And modified the following file to include the above test -

* src/test/regress/serial_schedule

Note: Any indexes built for the start expression of `substring` will
need to be rebuilt.

Note: A side effect found, because we use SQL's `substring`
implementation indirectly, is that there isn't a bounds check for
start values being less than 0. This needs to be addressed.

-jrg