fix: Make all Cypher log() calls use ln(), the natural logarithm
Notation: log10() refers to log base 10, ln() refers to log base e,
the natural logarithm.

The issue -

* Postgres SQL uses log10() for the log() function.

* Cypher uses ln() for the log() function.

* Our current implementation has a mix of log10() and ln() for
  Cypher queries using the log() function - based on type.

A decision was made to make our implementation of Cypher mirror the
standard. So, all Cypher log() function calls are to use ln() under
the hood.

To do this the following was changed -

* `parse_cypher_expr.c` was modified to transform all Cypher log()
  function calls to ln(), the Postgres internal function for ln().

* `cypher_func.sql` had regression tests added for the above change.

* `cypher_func.out` had the correct regression test output added
  for the above tests.

Note: This is a fix for AG-249

-jrg

(cherry picked from commit 28527415871d69c60cfe4d4929cad21a20f3e1d1)