fix: Remove jsonb type casts from pg_cast (AG-245)
This change removes jsonb type casts from pg_cast and adjusts the code
accordingly to provide those casts programmatically. See AG-245 for
more details.

In order to facilitate these changes, CoercionContext and TYPCATEGORY
fields are added to the CypherTypeCast structure. These additions
enable more granularity in the `ExecEvalCypherTypeCast` runtime
coercions.  This added granularity is needed to make the correct
casts at runtime.

The casts removed from pg_cast are as follows -

* Casts between JSONBOID and BOOLOID

* Casts from JSONBOID to INT4OID, INT8OID, FLOAT8OID, & NUMERICOID

These casts are programmatically added to -

* `coerce_expr` is rewritten to prioritize our JSONBOID casts over
  those of postgres, to create the CypherTypeCast for our casts, and
  to populate the CoercionContext and TYPCATEGORY fields.

* `eval_const_expressions_mutator` is adjusted to include the
  additions listed above in building new CypherTypeCast nodes.

* `ExecEvalCypherTypeCast` is adjusted to execute the new casts.

The `_copyCypherTypeCast`, `_equalCypherTypeCast`,
`_outCypherTypeCast`, and `_readCypherTypeCast` functions are modified
to support the added fields in CypherTypeCast.

The `makeDijkstraFrom` function is modified to use the new
`coerce_expr` function. Without this change it will throw an error due
to the missing type casts that are now incorporated in `coerce_expr`.

The following functions now use the new `coerce_cypher_arg_to_boolean`
function, which in turn uses the newer `coerce_expr` function -

* `transformCaseExpr` for CASE/WHEN clauses.

* `transformBoolExpr` for NOT/AND/OR expressions.

* `transformCypherWhere` for WHERE clauses.

The `jsonb_bool` function is now integrated in `ExecEvalCypherTypeCast`
and consequently, is removed.

The following regression tests are corrected to note the expected
failures. The original successes of these tests were due to type casts
that created side effects. So, now they fail - which is expected.

* `cypher_plpgsql.sql` and `cypher_plpgsql.out`.

* `cypher_substring.sql` and `cypher_substring.out`.

* `cypher_func.sql` (typo) and `cypher_func.out`

Note: Amended the commit to remove spaces, extra lines, adjust error
messaging in `ExecEvalCypherTypeCast`, to correct regression scripts'
output, and to fix one typo in a regression script.

-jrg