fix: ExecDijkstra crashes on FieldSelect exprs (AG-260)
This fix is to correct issue AG-260 which is due to FieldSelect
expressions that are unexpected in ExecDijkstra.

Please see commit d77f5bfc68e44c4136fa91e8db368cf8c40fdd09 for an
explaination of the issue. The fix is also similar, with the exception
of the need to only deal with the source node.

-jrg