fix: Could not determine collation (AG-262)
This fix is to correct AG-262, ERROR: could not determine which
collation to use for string comparison.

The issue is that the `transformCypherProjection` function casts the
targetList entries to jsonb after the groupClause is created. This
causes an issue with some of the collation values no longer being
accessible. Unfortunately, changing the position of the casting or
the groupClause won't correct the issue.

The fix is to copy over the collation values for all of the targetList
entries that were wrapped by the to_jsonb cast AND where the funccollid
value is not InvalidOid.

The function `repairTargetListCollations` is created to do just that.
It is added as the last operation in `transformCypherProjection`.

-jrg

(cherry picked from commit 9dbdab84a99ea2aeb8ad071cedb7499c4b7d449f)