fix: Dump output of ScalarArrayOpExpr for IN expression
When it is used for a graph constraint, the dump result of it looked
like below.

`CREATE CONSTRAINT ON label ASSERT ((prop) = ANY (ARRAY[...]));`

However, `ANY (ARRAY...` is not a valid expression in Cypher. So, the
function that is used to print out ScalarArrayOpExpr is modified
to print out the original expression of ScalarArrayOpExpr, and the
result now looks like below.

`CREATE CONSTRAINT ON label ASSERT ((prop) IN ([...]));`