fix: ExecNestLoopVLE crashes on match
ExecNestLoopVLE crashes on a match query where the vertices have no
lables AG-278.

This is due to an assert used to verify that chgParam is null in the
'previous context' logic.

The issue is that a chgParam is not being cleaned up after the inner
nested loop join fails to find any edges to follow. When this happens,
the inner plan never gets executed leaving an unprocessed chgParam. As
the inner plan is being reused, this chgParam effects the next
iteration across the inner plan.

When the previous context logic runs, the assert causes a crash.

The fix is to put code to clean up chgParam, instead of using assert,
for the following functions:

ExecPrevSeqScanContext, ExecPrevIndexScanContext, and
ExecPrevIndexOnlyScanContext.