fix: ExecNestLoopVLE hangs on some match queries
ExecNestLoopVLE hangs on some match queries AG-278.

Match queries for a path, that does not exist, between two vertices that
do exist, can cause ExecNestLoopVLE to go into an infinite loop.

This only happens when both the starting and ending vertices exist in
the graph, but there isn't a path between them. Additionally, the depth
of the search needs to be larger than the length of a dead end path from
the starting vertex.

The cause is an unresolved chgParam from the nested loop join looking
for edges on the last node of the path from the start vertex. This
leftover chgParam causes a rescan which results in the previous edge
being followed again.

The fix is to clear chgParam for all 'previous context' changes as we
will never revisit them. This was done in the following functions:

ExecPrevAppendContext and ExecPrevNestLoopContext.

(cherry picked from commit 7335ee511ed60b3014a36d667aca815d9b205159)