fix: enable_multiple_update regression test crashes
The cypher_dml.sql regression test crashes due to the
enable_multiple_update tests.

This crash has two parts -

The enable_multiple_update flag is used to disable (when set to false)
the pfreeing of hash table entries and to generate an error instead. So,
if it is false, we don't need to use pfree on the entries because we
won't be allowing the throwing out any previous values or replacing
them with any new updated values.

The other issue is due to the hash table persisting between invocations
of `ExecSetGraph`. The first calls to the hash table, via
`getVertexFinal` and `getEdgeFinal`, return the original slot value if
it found nothing - because that would be the most updated vertex or
edge. This meant that during the first run of the foreach loop in
`ExecSetGraph`, the slot value on entry to `ExecSetGraph` would also be
the finalValue for the updated vertex or edge in
`findAndReflectNewestValue`.

This logic was used to not pfree the original slot value - that we need
to not pfree.

However, the above assertion was incorrect because the hash table exists
between multiple invocations of `ExecSetGraph` for a query. So, it is
possible to find an updated value in that first run, causing it to pfree
the original slot value that shouldn't be freed. This is what caused the
crash.

The fix is to change the logic to avoid using pfree on the first
iteration of the foreach loop in `ExecSetGraph`.

This was done by adding a flag free_slot to `ExecSetGraph` to track the
first iteration of the foreach loop. This flag was then passed to
`findAndReflectNewestValue` which would use it to decide whether to
pfree the entry of not.

(cherry picked from commit 01597e13b9d11215788d293d9da779e20795c706)