fix: Wrong MemoryContext used for RTEs AG-281
Wrong MemoryContext used for RTEs in es_range_table.

Background-

ExecInitModifyGraph adds additional RTEs to the es_range_table. These
additional entries are added in the Executor context and then removed
in the Executor context in ExecEndModifyGraph. The es_range_table
itself contains entries added during the planning phase which uses a
derivative of the TopLevelContext.

Issue-

If the Executor phase improperly exits after these entries are added
but before they are deleted, the pointers will still remain in the
es_range_table. This will cause a segfault for a cached plan because the
Executor phase data structures pointed to, no longer exist when the cached
plan is run again. There are no issues for a non-cached plan.

Solution-

To correct this issue, in ExecInitModifyGraph, the RTEs added to the
es_range_table are added under the context used by the ModifyGraph plan.
Additionally, these RTEs are only added in once.

All of the other work in ExecInitModifyGraph is done under the Executor
context so that it can be cleaned up when the Executor exits or when
ExecEndModifyGraph runs.

To allow for proper handling of RTEs, ModifyGraph had two additional
variables added-

ert_base_index - To know where the plan's RTEs are added.
ert_rtes_added - To know how many RTEs have been added.

These variables are used, to know, when to add or remove RTEs and how to
reference the previously added RTEs.

These changes cause the RTEs added to the es_range_table to persist for
cached plans only. Additionally, all of the dependent ResultRelInfos get
cleaned up after the execution phase.

(cherry picked from commit 7d67dafcab3b70c1bac863f09678f8166f9f08c0)