refactor: VLE logic.
Previously, AgensGraph(NestLoopVLEJoin) was strongly combined with other
executors. This makes it difficult to identify PG's own code and AgensGraph
code, and it has a form that cannot be combined with other storage.

Due to this work, it will be easier to simplify merging upstream postgresql
source. also, help implement graph storage in the future.