fix: Includes necessary JOINs of vertices (#599)
Previously, a patch(3ced8823b026951549ef05f88cf4c479cb6080fb) was created to
remove the items that A thought were unnecessary joins. However, this returned
an exceptional case for the self-joining case without path(other words, Unnamed
path).

so, we rollback that partial of patch. and, we will refactor pattern match
logic for performance issue.